+++ ui-admin/src/study/kits/kitcollection/BarcodeScanner.tsx
